move python search configuration to its own module
