+++ python_search/cli.py
@@ -1,7 +1,8 @@
 import logging
 
 from python_search.apps.window_manager import WindowManager
-from python_search.config import ConfigurationLoader, PythonSearchConfiguration
+from python_search.configuration.configuration import PythonSearchConfiguration
+from python_search.configuration.loader import ConfigurationLoader
 from python_search.core_entities import Key
 from python_search.entry_runner import EntryRunner
 from python_search.environment import is_mac
