+++ python_search/config.py
@@ -3,16 +3,9 @@ Clients should depend on a configuration instance (config) rather than in the cl
 the class should only be used for type annotation.
 This way we can have multiple configs depending of the environment.
 """
-import datetime
-import logging
 import os
-from typing import List, Optional, Tuple, Literal
 
-from python_search.entries_group import EntriesGroup
 from python_search.environment import is_mac
-from python_search.features import PythonSearchFeaturesSupport
-
-
 
 
 class DataConfig:
@@ -51,149 +44,3 @@ class KafkaConfig:
 class RedisConfig:
     host = "127.0.0.1" if is_mac() else "host.docker.internal"
     port = 6379
-
-
-class PythonSearchConfiguration(EntriesGroup):
-    """
-    The main _configuration of Python Search
-    Everything to customize about the application is configurable via code through this class
-    """
-
-    APPLICATION_TITLE = "PythonSearchWindow"
-    commands: dict
-    simple_gui_theme = "SystemDefault1"
-    simple_gui_font_size = 14
-    _default_tags = None
-    tags_dependent_inserter_marks = None
-    _initialization_time = None
-    _default_text_editor = "vim"
-    _default_fzf_theme = None
-
-
-    def __init__(
-        self,
-        *,
-        entries: Optional[dict] = None,
-        entries_groups: Optional[List[EntriesGroup]] = None,
-        supported_features: Optional[PythonSearchFeaturesSupport] = None,
-        default_tags=None,
-        tags_dependent_inserter_marks: Optional[dict[str, Tuple[str, str]]] = None,
-        default_text_editor: Optional[str] = None,
-        default_fzf_theme: Optional[Literal["light"]] = None,
-        custom_window_size: Optional[Tuple[int, int]] = None,
-    ):
-        """
-
-        :param entries:
-        :param entries_groups:
-        :param supported_features:
-        :param default_tags:
-        :param tags_dependent_inserter_marks:
-        :param default_text_editor:
-        :param default_fzf_theme: the theme to use for fzf
-        :param custom_window_size: the size of the fzf window
-        """
-        if entries:
-            self.commands = entries
-
-        if entries_groups:
-            self.aggregate_commands(entries_groups)
-
-        if supported_features:
-            self.supported_features: PythonSearchFeaturesSupport = supported_features
-        else:
-            self.supported_features: PythonSearchFeaturesSupport = (
-                PythonSearchFeaturesSupport.default()
-            )
-
-        if default_tags:
-            self._default_tags = default_tags
-
-        self.tags_dependent_inserter_marks = tags_dependent_inserter_marks
-
-        self._initialization_time = datetime.datetime.now()
-        self._default_text_editor = default_text_editor
-        self._default_fzf_theme = default_fzf_theme
-        if custom_window_size:
-            self._custom_window_size = custom_window_size
-
-
-
-    def get_text_editor(self):
-        return self._default_text_editor
-
-    def get_default_tags(self):
-        return self._default_tags
-
-    def get_fzf_theme(self):
-        return self._default_fzf_theme
-
-    def get_window_size(self):
-        if hasattr(self, "_custom_window_size"):
-            return self._custom_window_size
-
-
-class ConfigurationLoader:
-    """
-    Loads the application from the environment.py
-    """
-
-    def load_config(self) -> PythonSearchConfiguration:
-
-        folder = self.get_project_root()
-        config_location = f"{folder}/entries_main.py"
-
-        if not os.path.exists(config_location):
-            raise Exception(f"Could not find config file {config_location}")
-
-        import sys
-
-        if folder not in sys.path:
-            sys.path.insert(0, folder)
-        import copy
-        from entries_main import config
-
-        return copy.deepcopy(config)
-
-    def reload(self):
-        """
-        reload _entries for when we change it
-        """
-        import importlib
-
-        import entries_main
-
-        importlib.reload(entries_main)
-
-        import copy
-
-        import entries_main
-
-        return copy.deepcopy(entries_main.config)
-
-    def get_project_root(self):
-        env_name = "PS_ENTRIES_HOME"
-        current_project_location = (
-            os.environ["HOME"] + "/.config/python_search/current_project"
-        )
-
-        folder = None
-
-        if env_name in os.environ:
-            logging.debug(
-                f"Env exists and takes precedence: {env_name}={os.environ[env_name]}"
-            )
-            folder = os.environ[env_name]
-
-        if os.path.isfile(current_project_location):
-            with open(current_project_location) as f:
-                folder = f.readlines()[0].strip()
-
-        if not folder:
-            raise Exception(
-                f"Either {current_project_location} or {env_name} must be set to find _entries"
-            )
-        return folder
-
-    def load_entries(self):
-        return self.load_config().commands
