+++ python_search/search/next_item_predictor/inference/inference.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 import os
 from typing import Any, List, Optional
 
-from python_search.config import ConfigurationLoader, PythonSearchConfiguration
+from python_search.configuration.configuration import PythonSearchConfiguration
+from python_search.configuration.loader import ConfigurationLoader
 from python_search.infrastructure.performance import timeit
 from python_search.logger import setup_inference_logger
 from python_search.search.next_item_predictor.inference.input import ModelInput
@@ -25,13 +26,8 @@ class Inference:
     ):
 
         self.debug = os.getenv("DEBUG", False)
-        if "FORCE_RUN_ID" in os.environ:
-            self.run_id = os.environ["FORCE_RUN_ID"]
 
-        if model:
-            logger.info("Using custom passed model")
-        else:
-            logger.info("Next item predictor using run id: " + self.run_id)
+
         configuration = (
             configuration if configuration else ConfigurationLoader().load_config()
         )
@@ -39,6 +35,12 @@ class Inference:
         self._model = NextItemModelV1()
 
         self.run_id = run_id if run_id else self._model.get_run_id()
+
+        if model:
+            logger.info("Using custom passed model")
+        else:
+            logger.info("Next item predictor using run id: " + self.run_id)
+
         try:
             self._mlflow_model = model if model else self._model.load_mlflow_model(run_id=self.run_id)
         except Exception as e:
