+++ python_search/search/next_item_predictor/offline_evaluation.py
@@ -1,6 +1,6 @@
 from pyspark.sql import DataFrame
 
-from python_search.config import ConfigurationLoader
+from python_search.configuration.loader import ConfigurationLoader
 from python_search.search.next_item_predictor.inference.inference import Inference
 from python_search.search.next_item_predictor.inference.input import ModelInput
