+++ python_search/search_ui/fzf_kitty.py
@@ -1,7 +1,7 @@
 import os
 
 from python_search.apps.terminal import Terminal
-from python_search.config import PythonSearchConfiguration
+from python_search.configuration.configuration import PythonSearchConfiguration
 from python_search.environment import is_mac
