+++ python_search/web_api.py
@@ -12,7 +12,7 @@ from python_search.entry_type.classifier_inference import (
 from python_search.events.latest_used_entries import RecentKeys
 from python_search.events.run_performed import RunPerformed
 from python_search.events.run_performed.writer import RunPerformedWriter
-from python_search.config import ConfigurationLoader
+from python_search.configuration.loader import ConfigurationLoader
 from python_search.search.search import Search
 
 import pyroscope
